+++ test/hoomd_script/test_multiple_contexts.py
@@ -6,13 +6,13 @@ context.initialize()
 import unittest
 import os
 
-# unit test to run a simple polymer system with pair and bond potentials
+# test multiple contexts with a simple run
 class multi_context(unittest.TestCase):
     def test_run(self):
-        self.c1 = context.SimulationContext()
-        self.c2 = context.SimulationContext()
+        c1 = context.SimulationContext()
+        c2 = context.SimulationContext()
 
-        with self.c1:
+        with c1:
             init.create_random(N=2000, phi_p=0.2)
             lj = pair.lj(r_cut=3.0)
             lj.pair_coeff.set('A', 'A', epsilon=1.0, sigma=1.0)
@@ -20,7 +20,7 @@ class multi_context(unittest.TestCase):
             integrate.mode_standard(dt=0.005)
             integrate.nvt(group=group.all(), T=1.2, tau=0.5)
 
-        with self.c2:
+        with c2:
             init.create_random(N=1000, phi_p=0.02)
             lj = pair.lj(r_cut=3.0)
             lj.pair_coeff.set('A', 'A', epsilon=1.0, sigma=1.0)
@@ -28,11 +28,30 @@ class multi_context(unittest.TestCase):
             integrate.mode_standard(dt=0.005)
             integrate.nvt(group=group.all(), T=1.2, tau=0.5)
 
-        with self.c1:
+        with c1:
             run(10)
 
-        with self.c2:
+        with c2:
             run(10)
+
+    def test_replace(self):
+        old = context.current;
+        c1 = context.SimulationContext()
+
+        with c1:
+            pass
+
+        # current context should be reset to the old
+        print(context.current)
+        print(old)
+        print(c1)
+        self.assertTrue(context.current is old);
+
+    def test_set_current(self):
+        c1 = context.SimulationContext()
+        c1.set_current();
+        self.assertTrue(context.current is c1);
 
 if __name__ == '__main__':
     unittest.main(argv = ['test.py', '-v'])
+